Add files via upload
I have implemented doubly-linked-list to make a To-do-List. It is a menu driven program that allows users to add tasks as many as they want, mark them as completed, and remove a specific task or all the completed tasks. The program also allows the user to print the list of tasks in both orders (forward and reverse).